Ageing error software

R package to support the Punt et al. (2008) ageing error model

References

Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008. Quantifying age-reading error for use in fisheries stock assessments, with application to species in Australias southern and eastern scalefish and shark fishery. Canadian Journal of Fisheries and Aquatic Sciences 65: 1991-2005.

See also

Author

Maintainer: Ian G. Taylor ian.taylor@noaa.gov (ORCID)

Authors:

Andre E. Punt (ORCID)

Kelli F. Johnson (ORCID)

Paul Burch (ORCID)

Other contributors:

James T. Thorson (ORCID) [contributor, copyright holder]

Ian J. Stewart [contributor]

Melissa A. Haltuch (ORCID) [contributor]